refactor: move fullscreen button to common components
