+++ src/components/FullscreenButton/index.tsx
@@ -1,0 +1,1 @@
+export { FullscreenButton } from './FullscreenButton';
